SWServer::removeClientServiceWorkerRegistration should not log an error if there is no registration https://bugs.webkit.org/show_bug.cgi?id=183248
Patch by Youenn Fablet <youenn@apple.com> on 2018-03-01
Reviewed by Chris Dumez.

No change of behavior.
Removed the error logging.
As can be seen from running api and layout tests, a web process can ask a registration to be unregistered
and by the time it goes to the StorageProcess, the registration is already cleared.

* workers/service/server/SWServer.cpp:
(WebCore::SWServer::removeClientServiceWorkerRegistration):

Identifier: 198936@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229149 268f45cc-cd09-0410-ab3c-d52691b4dbfc