Delete incorrect version of clampTo() function from SVGToOTFFontConversion.cpp https://bugs.webkit.org/show_bug.cgi?id=183165
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2018-03-03
Reviewed by Darin Adler.

Source/WebCore:

Some of the calls in SVGToOTFFontConversion.cpp were directed to the function
in MathExtras.h while the rest were directed to this local static function.
There should not be two versions with the same name while they are supposed
to do the same thing. Besides, the local version does not work correctly
if the type of the argument's max limit is less the max limit of the returned
type, e.g. char -> uint16_t.

* svg/SVGToOTFFontConversion.cpp:
(WebCore::clampTo): Deleted.

LayoutTests:

Re-baseline the results of these two tests since the local version of
clampTo() was not returning the expected results always.

* platform/ios/svg/W3C-SVG-1.1/text-altglyph-01-b-expected.txt:
* platform/ios/svg/text/text-altglyph-01-b-expected.txt:

Identifier: 198983@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229202 268f45cc-cd09-0410-ab3c-d52691b4dbfc