Changing link element rel attribute from preload to stylesheet should succeed loading the stylesheet https://bugs.webkit.org/show_bug.cgi?id=183601 <rdar://problem/38309441>
Reviewed by Antti Koivisto.

Source/WebCore:

Test: http/wpt/preload/change-link-rel-attribute.html

* loader/LinkPreloadResourceClients.h:
(WebCore::LinkPreloadResourceClient::clearResource): Remove the call to CachedResource::cancelLoad.
This call is expected to be called by ResourceLoader when cancelling the load from below CachedResource.
* loader/cache/CachedResource.cpp:
(WebCore::CachedResource::allClientsRemoved): In case of preload,
cancel the load if not finished when there is no more client attached to it.
* loader/cache/CachedResource.h:

LayoutTests:

* http/wpt/preload/change-link-rel-attribute-expected.txt: Added.
* http/wpt/preload/change-link-rel-attribute.html: Added.
* http/wpt/preload/resources/style.css: Added.
(body):


Identifier: 199270@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229585 268f45cc-cd09-0410-ab3c-d52691b4dbfc