Make Fullscreen API an Experimental Feature https://bugs.webkit.org/show_bug.cgi?id=183662
Reviewed by Jon Lee.

Source/WebCore:

The Fullscreen APIs should be guarded by a Setting, so they are not present in the DOM when the
Experimental Feature is disabled.

* dom/Document.idl:
* dom/Element.idl:

Source/WebKit:

Add the ability to have a conditional for whether an Experimental Feature is "visible", separate from
whether the preference is available, for definining platform-specific Experimental Features.

* Scripts/GeneratePreferences.rb:
* Scripts/PreferencesTemplates/WebPreferencesExperimentalFeatures.cpp.erb:
* Shared/WebPreferences.yaml:


Identifier: 199347@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@229680 268f45cc-cd09-0410-ab3c-d52691b4dbfc