[Web Animations] Comparing two matching FramesTimingFunction objects fails https://bugs.webkit.org/show_bug.cgi?id=184078
Reviewed by Jon Lee.

LayoutTests/imported/w3c:

Mark one progression for the frames() timing function test with the CSS Animations and CSS Transitions as Web Animations flag.

* web-platform-tests/css-timing-1/frames-timing-functions-output-expected.txt:
* web-platform-tests/css-timing-1/frames-timing-functions-output.html:

Source/WebCore:

Fix a stupid mistake in the == operator for FramesTimingFunction. Without a correct check, we would mistakenly interrupt and
restart any transition with a frames() timing function when we check that the Animation objects are similar.

* platform/animation/TimingFunction.h:

Identifier: 199663@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@230031 268f45cc-cd09-0410-ab3c-d52691b4dbfc